Some preliminary work on splitting render-vk
